+++ CHANGELOG.md
@@ -11,6 +11,7 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 * Fix KML export for events with no forms.
 * Fix bug in MGRS display for lines and polygons.
 * Fix username/password error message on invalid login when account lock is disabled.
+* Fix bug in observation edit where initial timezone format was not being picked up.
 
 ## [5.2.0](https://github.com/ngageoint/mage-server/releases/tag/5.2.0) (09-17-2018)